# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the ephemeral secret-sharing utility
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RoomInput {
  /**
     * @minimum 1
     * @maximum 120
     */
  lifetimeMinutes?: number;
}

export interface ParticipantInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name?: string;
}

export interface SecretInput {
  hostToken: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  value: string;
}

export interface SecretUpdate {
  hostToken: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  value: string;
}

export interface SecretDeleteRequest {
  hostToken: string;
}

export interface Secret {
  id: string;
  label: string;
  value: string;
  createdAt: string;
}

export interface Room {
  code: string;
  expiresAt: string;
  participantCount: number;
  secrets: Secret[];
}

export type RoomSessionRole = typeof RoomSessionRole[keyof typeof RoomSessionRole];


export const RoomSessionRole = {
  host: 'host',
  participant: 'participant',
} as const;

export type RoomSession = Room & ({
  role: RoomSessionRole;
  participantId: string;
  /** @nullable */
  hostToken?: string | null;
});

export interface ErrorResponse {
  error: string;
}

/**
 * Room was not found or has expired
 */
export type NotFoundResponse = ErrorResponse;