# lib/api-zod/src/generated/types/roomSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the ephemeral secret-sharing utility
 * OpenAPI spec version: 0.1.0
 */
import type { Room } from './room';
import type { RoomSessionRole } from './roomSessionRole';

export type RoomSession = Room & ({
  role: RoomSessionRole;
  participantId: string;
  /** @nullable */
  hostToken?: string | null;
});