# lib/api-zod/src/generated/types/secret.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the ephemeral secret-sharing utility
 * OpenAPI spec version: 0.1.0
 */

export interface Secret {
  id: string;
  label: string;
  value: string;
  createdAt: Date;
}