# lib/api-zod/src/generated/types/secretUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the ephemeral secret-sharing utility
 * OpenAPI spec version: 0.1.0
 */

export interface SecretUpdate {
  hostToken: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  value: string;
}